Put error handling in error file
