+++ src/core/error.ts
@@ -59,3 +59,22 @@ export class HttpError extends Error {
         this.name = "HttpError";
     }
 }
+
+function isTelegramError(
+    err: unknown,
+): err is { status: string; statusText: string } {
+    return (
+        typeof err === "object" &&
+        err !== null &&
+        "status" in err &&
+        "statusText" in err
+    );
+}
+export function toHttpError(method: string, sensitiveLogs: boolean) {
+    return (err: unknown) => {
+        let msg = `Network request for '${method}' failed!`;
+        if (isTelegramError(err)) msg += ` (${err.status}: ${err.statusText})`;
+        if (sensitiveLogs && err instanceof Error) msg += ` ${err.message}`;
+        throw new HttpError(msg, err);
+    };
+}
